Remote client IP Addresses for BlackList

+++ AuctionSystem.Server/Services/UserService.cs
@@ -46,7 +46,7 @@ namespace AuctionSystem.Server.Services
                     };
                     BlackListIp blackListIp = new BlackListIp()
                     {
-                        Ipaddress = (int)model.IPAddress,                        
+                        Ipaddress = model.IPAddress,                        
                         Active = true,
                         CreatedOn = DateTime.Now,
                         Id = Guid.NewGuid()
